yapf formatting

+++ baselines/shampoo/jax/distributed_shampoo.py
@@ -595,8 +595,8 @@ def matrix_inverse_pth_root(
 
   if padding_start is not None:
     # Zero out padding in identity as well for convergence checks.
-    ix = (jnp.arange(matrix_size, dtype=jnp.int32) < padding_start).astype(
-        matrix.dtype)
+    ix = (jnp.arange(matrix_size, dtype=jnp.int32)
+          < padding_start).astype(matrix.dtype)
     matrix *= ix[jnp.newaxis, :]
     matrix *= ix[:, jnp.newaxis]
     identity *= ix
@@ -815,8 +815,8 @@ def matrix_inverse_pth_root_eigh(
   alpha = jnp.asarray(-1.0 / p, _MAT_INV_PTH_ROOT_DTYPE)
   identity = jnp.eye(matrix_size, dtype=_MAT_INV_PTH_ROOT_DTYPE)
   if padding_start is not None:
-    ix = (jnp.arange(matrix_size, dtype=jnp.int32) < padding_start).astype(
-        matrix.dtype)
+    ix = (jnp.arange(matrix_size, dtype=jnp.int32)
+          < padding_start).astype(matrix.dtype)
     matrix *= ix[jnp.newaxis, :]
     matrix *= ix[:, jnp.newaxis]
     identity *= ix
@@ -1923,8 +1923,8 @@ def distributed_shampoo(
     errors = metrics.inverse_pth_root_errors
     errors = errors.reshape((-1, 1, 1))
     predicate = jnp.logical_or(
-        jnp.isnan(errors),
-        errors >= inverse_failure_threshold).astype(new_preconditioners.dtype)
+        jnp.isnan(errors), errors
+        >= inverse_failure_threshold).astype(new_preconditioners.dtype)
     # TODO(rohananil): Check for numerical instabilities.
     new_conditional_preconditioners = (
         predicate * global_stats.preconditioners +
@@ -2190,9 +2190,8 @@ def distributed_shampoo(
       new_preconditioners_flat.append(
           _select_preconditioner(error, p[:shape[0], :shape[1]], prev_p))
 
-    assert len(states) == (len(
-        num_statistics_per_state),
-        f"{len(states)} vs {len(num_statistics_per_state)}")
+    assert len(states) == (len(num_statistics_per_state),
+                           f"{len(states)} vs {len(num_statistics_per_state)}")
     assert len(new_preconditioners_flat) == num_statistics
     assert len(new_errors_flat) == len(packed_statistics), (
         len(new_errors_flat), len(packed_statistics))
@@ -2285,9 +2284,8 @@ def distributed_shampoo(
         original_shapes.extend(original_shapes_for_state)
 
     # if quantized_dtype == jnp.float32:
-    assert len(states) == (len(
-        num_statistics_per_state),
-        f"{len(states)} vs {len(num_statistics_per_state)}")
+    assert len(states) == (len(num_statistics_per_state),
+                           f"{len(states)} vs {len(num_statistics_per_state)}")
     return _pmap_compute_preconditioners(states,
                                          step,
                                          statistics,
